+++ fbw-common/src/typings/fs-base-ui/html_ui/JS/Flight.d.ts
@@ -1,3 +1,4 @@
+/// <reference path="../../../units.d.ts" />
 /// <reference path="./Types.d.ts" />
 /// <reference path="./common.d.ts" />
