+++ fbw-common/src/typings/index.d.ts
@@ -4,4 +4,4 @@ export * from './asobo-vcockpits-instruments-airliners';
 export * from './flybywire-vcockpits-instruments';
 export * from './fs-base-ui';
 export * from './types';
-export * from './native';
+export * from './units';
